Add css for 02-class-id-selectors
I've used a different kind of approach.
I decided to group the odd number elements and specified a another
declaration for adjusting the font-size of the third element.

I did not use  multiple classes for the third element.

I've also used the three different methods for specifying a color value
using the `hsl()`, `rgb(), and the hex values.