Fix css for 01-margin-and-padding-1 exercise
I added a padding of 32px for both Div One and Div Three.
Div one has 12px margin between any elements on the page.
I added a 48px of margin-bottom for Div Two.
And, Div Three is aligned to the right.

What I did is, calculated the total width of Div Two and Div Three.

And subtract 270px total width of Div Three to total width of Div Two of
600px. The result it 330px.

.three has a total width of 270px
.two has a total width of 600px

I then, set 330px of margin-left in Div Three.