Fix css for 02-margin-padding-2 exercise
8px of padding is set to .card to achieve a space between the edge of
the card and its content(the blue sections).

8px of margin-bottom is set to .title and .content and .title uses a
16px of font-size.

An 8px of padding is added in .title.

A padding: 16px 8px; is added in .content section. And padding of 8px
and text-align: center; is added in the .button-container.

Big Button is centered in its own line using the display: block;
And margin: auto; padding: 8px 24px; is added.