Fix flex for 02-flex-header
There is a space between all items and the edge of the header.
The Logo is centered vertically and horizontally.
list-items are horizontal, and are centered vertically inside the header.
left-links and right-links are pushed all the way to the left and right, and stay at the edge of the header when the page is resized.

As for when resizing, I've added the property flex-wrap: wrap;