Fix flex for 03-flex-header-2
Everything is centered vertically inside the header.
There is 8px space between everything and the edge of the header.
Items are arranged horizontally as seen in the outcome image.
There is 16px between each item on both sides of the header.
flex is used to arrange everything.

So far, I have completed the self-check. I've learned a lot here, being
able to create navbar like, using flexbox is cool.

I created two (2) more flex container to group the left and right items.
With that, I've added a unnecessary `justify-content` in the flex
childs. The separation of right and left containers is achieved by the
`justify-content: space-between;`. in the parent flex.