Fix flex for 05-flex-modal
The blue icon is aligned to the left.
There is equal space on either side of the icon (the gaps between the icon and the edge of the card, and the icon and the text, are the same).
There is padding around the edge of the modal.
The header, text, and buttons are aligned with each other.
The header is bold and a slightly larger text-size than the text.
The close button is vertically aligned with the header, and aligned in the top-right of the card.