Fix flex for 06-flex-layout
The header is at the top of the page, the footer is at the bottom, and they stay in place if you resize your screen.
The header and footer have padding.
The links in the header and footer are pushed to either side.
There is space between the links in the header and footer.
The footer has a light gray background (`#eeeeee`).
The logo, input and buttons are centered in the screen.
The buttons have an appropriate amount of padding.
There is space between the logo, input and buttons.

Overall the self-checks are achieved, I learned a lot in this activity.
The gap property is so useful when you need to space out every flex item
in a container, like for example in the .content container, adding the
gap, allocates a gap for (logo, input, and buttons) every flex item.

Another thing I learned even though, I have actually did this in the
past activities. Is that by setting the flex-direction for the container
is a must, depends on the design how should the flex items interact with
each other. Therefore by using the flex-direction: column, in the body
element. It will stack nicely the .header, .content, and .footer from
top to bottom, which is very useful again...

Lastly, the removal of links by using text-docoration: none is very
cool.