Initial fix for 07-flex-layout-2
These are the self checks that I think I completed it.
For some reason, I am not sure of the method I used when aligning the
cards in the main content.

I am still confused when setting properties like flex-grow
flex-basis and flex-shrink.

The header text is size 32px and weight 900.
The header text is vertically centered and 16px from the edge of the screen.
The footer is pushed to the bottom of the screen (the footer may go _below_ the bottom of the screen if the content of the 'cards' section overflows and/or if your screen is shorter).
The footer text is centered horizontally and vertically.
The sidebar and cards take up all available space above the footer.
The sidebar is 300px wide (and it doesn't shrink).
The sidebar links are size 24px, are white, and do not have the underline text decoration.
The sidebar has 16px padding.
There is 48px padding around the 'cards' section.
The cards are arranged horizontally, but wrap to multiple lines when they run out of room on the page.